backend/requirements.txt
# Backend Dependencies for Conversational RAG System

# FastAPI and server
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.9,<1.0.0

# Pydantic for validation
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# LangChain - using latest compatible versions
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0

# Embeddings
sentence-transformers>=2.2.0

# Vector store
faiss-cpu>=1.7.0

# Document parsing
pypdf>=4.0.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
